% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mq_raw_TMT}
\alias{mq_raw_TMT}
\title{Wine reviews for 51 countries.}
\format{A data frame with 280901 rows and 4 variables:
\describe{
\item{country}{The country that the wine is from.}
\item{points}{The number of points rated by WineEnthusiast
on a scale of 1–100.}
\item{price}{The cost for a bottle of the wine.}
\item{winery}{The winery that made the wine.}
}}
\source{
\url{https://www.kaggle.com/zynicide/wine-reviews}
}
\usage{
mq_raw_TMT
}
\description{
A dataset containing wine reviews.
}
\keyword{datasets}
